refactor: rename `dist` to `disturbance`
`dist` is a special term for most tool filters. `pytest` was skipping our `dist`urbance tests because it
thought it was a build artifact.
