+++ pipelines/disturbance/validate_zonal_statistics.py
